+++ stac-example/src/main/scala/geotrellis/server/ogc/conf/OgcSourceConf.scala
@@ -30,6 +30,7 @@ import geotrellis.stac.raster.StacCollectionSource
 import pureconfig.ConfigReader
 
 import scala.util.Try
+import scala.util.matching.Regex
 
 // This sumtype corresponds to the in-config representation of a source
 sealed trait OgcSourceConf {
@@ -45,7 +46,7 @@ case class StacSourceConf(
   collection: Option[String],
   title: String,
   source: String,
-  asset: String,
+  asset: Regex,
   assetLimit: Option[NonNegInt],
   pageLimit: Option[NonNegInt],
   defaultStyle: Option[String],